services/web/requirements.txt
Werkzeug~=3.1.3
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
Flask-Migrate==4.1.0
Flask-Login==0.6.3
PyMySQL==1.1.1
redis>=5.0.0